refactor: cleanup job macro impl
