+++ crates/macros/blueprint-proc-macro/src/job/evm.rs
@@ -1,10 +1,50 @@
+use super::args::EventListenerArgs;
+use super::declared_params_to_field_types;
+use crate::shared::{get_non_job_arguments, get_return_type_wrapper};
 use indexmap::IndexMap;
 use proc_macro2::{Span, TokenStream};
 use quote::{format_ident, quote};
-use syn::{Ident, Type};
+use syn::parse::{Parse, ParseStream};
+use syn::{Ident, Token, Type};
 
-use crate::job::{declared_params_to_field_types, EventListenerArgs};
-use crate::shared::{get_non_job_arguments, get_return_type_wrapper};
+/// Defines custom keywords for defining Job arguments
+mod kw {
+    syn::custom_keyword!(instance);
+    syn::custom_keyword!(abi);
+}
+
+pub(crate) struct EvmArgs {
+    pub instance: Option<Ident>,
+    pub abi: Option<Type>,
+}
+
+impl Parse for EvmArgs {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        let content;
+        syn::parenthesized!(content in input);
+
+        let mut instance = None;
+        let mut abi = None;
+
+        while !content.is_empty() {
+            if content.peek(kw::instance) {
+                let _ = content.parse::<kw::instance>()?;
+                let _ = content.parse::<Token![=]>()?;
+                instance = Some(content.parse::<Ident>()?);
+            } else if content.peek(Token![,]) {
+                let _ = content.parse::<Token![,]>()?;
+            } else if content.peek(kw::abi) {
+                let _ = content.parse::<kw::abi>()?;
+                let _ = content.parse::<Token![=]>()?;
+                abi = Some(content.parse::<Type>()?);
+            } else {
+                return Err(content.error("Unexpected token"));
+            }
+        }
+
+        Ok(EvmArgs { instance, abi })
+    }
+}
 
 pub(crate) fn get_evm_instance_data(
     event_handler: &EventListenerArgs,
