--- crates/macros/blueprint-proc-macro/src/special_impls/mod.rs
@@ -1,2 +1,0 @@
-pub mod evm;
-pub mod tangle;
